+++ src/lib/utils/index.ts
@@ -12,5 +12,6 @@ export * from "./redo";
 export * from "./scroll";
 export * from "./textLine";
 export * from "./date";
+export * from "./validate";
 export * from "./parser";
 export * from "./option";